+++ app/Models/Imovel.php
@@ -8,7 +8,7 @@ use App\Models\Inquilinos;
 class Imovel extends Model
 {
     use HasFactory;
-
+   
     protected $table = 'imoveis';
 
     protected $fillable = [