+++ database/migrations/2024_09_19_135326_create_inquilinos_table.php
@@ -19,6 +19,7 @@ return new class extends Migration
             $table->date('data_inicio_contrato')->nullable();
             $table->date('data_fim_contrato')->nullable();
             $table->string('status');
+            $table->string('contrato_pdf');
             $table->string('pagamento_recente')->nullable();
         });
     }